backend/requirements.txt
fastapi
uvicorn
pydantic
pem
python-jose
pycryptodome
python-multipart
